refactor(scanner): extract strategy/patterns.py — engulfings, divergences, triggers, score_label (#225 PR1)
